Support pre v0.10 snaps

+++ pkg/controller/backup_deletion_controller.go
@@ -245,25 +245,47 @@ func (c *backupDeletionController) processRequest(req *v1.DeleteBackupRequest) e
 
 	if backupStore != nil {
 		log.Info("Removing PV snapshots")
-		if snapshots, err := backupStore.GetBackupVolumeSnapshots(backup.Name); err != nil {
-			errs = append(errs, errors.Wrap(err, "error getting backup's volume snapshots").Error())
-		} else {
-			blockStores := make(map[string]cloudprovider.BlockStore)
-
-			for _, snapshot := range snapshots {
-				log.WithField("providerSnapshotID", snapshot.Status.ProviderSnapshotID).Info("Removing snapshot associated with backup")
-
-				blockStore, ok := blockStores[snapshot.Spec.Location]
-				if !ok {
-					if blockStore, err = blockStoreForSnapshotLocation(backup.Namespace, snapshot.Spec.Location, c.snapshotLocationLister, pluginManager); err != nil {
-						errs = append(errs, err.Error())
-						continue
+		if len(backup.Status.VolumeBackups) > 0 {
+			// pre-v0.10 backup
+			locations, err := c.snapshotLocationLister.VolumeSnapshotLocations(backup.Namespace).List(labels.Everything())
+			if err != nil {
+				errs = append(errs, errors.Wrap(err, "error listing volume snapshot locations").Error())
+			} else if len(locations) != 1 {
+				errs = append(errs, errors.Errorf("unable to delete pre-v0.10 volume snapshots because exactly one volume snapshot location must exist, got %d", len(locations)).Error())
+			} else {
+				blockStore, err := blockStoreForSnapshotLocation(backup.Namespace, locations[0].Name, c.snapshotLocationLister, pluginManager)
+				if err != nil {
+					errs = append(errs, err.Error())
+				} else {
+					for _, snapshot := range backup.Status.VolumeBackups {
+						if err := blockStore.DeleteSnapshot(snapshot.SnapshotID); err != nil {
+							errs = append(errs, errors.Wrapf(err, "error deleting snapshot %s", snapshot.SnapshotID).Error())
+						}
 					}
-					blockStores[snapshot.Spec.Location] = blockStore
 				}
+			}
+		} else {
+			// v0.10+ backup
+			if snapshots, err := backupStore.GetBackupVolumeSnapshots(backup.Name); err != nil {
+				errs = append(errs, errors.Wrap(err, "error getting backup's volume snapshots").Error())
+			} else {
+				blockStores := make(map[string]cloudprovider.BlockStore)
 
-				if err := blockStore.DeleteSnapshot(snapshot.Status.ProviderSnapshotID); err != nil {
-					errs = append(errs, errors.Wrapf(err, "error deleting snapshot %s", snapshot.Status.ProviderSnapshotID).Error())
+				for _, snapshot := range snapshots {
+					log.WithField("providerSnapshotID", snapshot.Status.ProviderSnapshotID).Info("Removing snapshot associated with backup")
+
+					blockStore, ok := blockStores[snapshot.Spec.Location]
+					if !ok {
+						if blockStore, err = blockStoreForSnapshotLocation(backup.Namespace, snapshot.Spec.Location, c.snapshotLocationLister, pluginManager); err != nil {
+							errs = append(errs, err.Error())
+							continue
+						}
+						blockStores[snapshot.Spec.Location] = blockStore
+					}
+
+					if err := blockStore.DeleteSnapshot(snapshot.Status.ProviderSnapshotID); err != nil {
+						errs = append(errs, errors.Wrapf(err, "error deleting snapshot %s", snapshot.Status.ProviderSnapshotID).Error())
+					}
 				}
 			}
 		}
